Explicit exception chaining and enhanced reporting.
Converting all exceptions to OsbsException can lose information about
the context of the original exception. It would be nice to be able to
show that context.

This happens implicitly with Python 3, but to aid reporting with
Python 2 we can do it explicitly.

This is only activated in Python 2, by detecting whether the
__context__ attribute is set.