Fake API: framework changes to allow responses to change over time
The URL to response map has been moved into the Connection object, as
well as the custom callbacks referenced by it and the lookup function,
get_definition_for().

This lookup function now also returns the key in the response map,
allowing a custom callback to modify the response for that URL the
next time it is requested.