Pause/resume scheduling of new builds
This admin function can be useful when maintenance needs to be
performed on external systems such as Pulp, a Docker registry, NFS
server, etc.

For 'pause-builds' the resource quota for the number of running pods
is temporarily set to 0, preventing new builds from being scheduled,
and running builds are left to finish.

For 'resume-builds' the resource quota of 0 running pods is lifted,
leaving things the way they were before the pause.