Improve performance for `list-builds --running`
Previously, all builds were fetched and then each build's status was
examined.

Now, the fieldSelector is set to match any unfinished status in the builds
query when the --running option is passed to the CLI tool. This means
only new/pending/running builds will be returned.

API changes:
- OSBS.list_builds() now accepts a 'field_selector' keyword argument

This also brings us closer to an implementation for #252.