Coordinate plugins for orchestrator/worker builds
This commit introduces the concept of an 'arrangement version'. For
orchestrator builds, the plugins run in the orchestrator build and those
run in the worker build must be coordinated.  Changes to this
arrangement will be made by defining a new arrangement version and
creating new 'inner' JSON templates for the orchestrator and worker
builds. The DEFAULT_ARRANGEMENT_VERSION constant will be incremented to
match this new version.

The osbs.conf configuration file now has an arrangement_version keyword,
which defaults to DEFAULT_ARRANGEMENT_VERSION.

The create_orchestrator_build() method takes an optional keyword
parameter 'arrangement_version', which defaults to the value provided
from configuration. This is used to select the orchestrator_inner:n.json
template to use, as well as being provided to the orchestrate_build
plugin as part of its build_kwargs parameter.

The create_worker_build() method has a new required
'arrangement_version' parameter. This is used to select the
worker_inner:n.json template to use.

When a requested arrangement version is not present (the inner template
file does not exist) for either an orchestrator or worker build,
OsbsValidationException is raised.

This is made possible by avoiding nested calls to @osbsapi-wrapped
methods. The create_prod_build method is renamed to
_do_create_prod_build (which is not @osbsapi-wrapped), and
create_*_build methods all use this.

Signed-off-by: Tim Waugh <twaugh@redhat.com>