Make scratch build and output names unique
Use random 5 digit number to make the name of the output
unique (in addition to the timestamp already used).

This should ensure, that scratch builds started in the same time
will not be tagged with the same name in the registry.

Reuse this unique stamp (5 random digits and timestamp)
for scratch build names, in order to avoid build name collision.

Resolves #483.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>