support old setuptools on el7
The ancient setuptools in el7 behaves differently from more recent
setuptools, which results in unit test failures when running on an el7
host.

A quick illustration of the differences:

  # Python 2.7.13 (fedora)
  >>> from pkg_resources import parse_version
  >>> parse_version('1.0.1')
  <Version('1.0.1')>
  >>> str(parse_version('1.0.1'))
  '1.0.1'

  # Python 2.7.5 (rhel 7)
  >>> from pkg_resources import parse_version
  >>> parse_version('1.0.1')
  ('00000001', '00000000', '00000001', '*final')
  >>> str(parse_version('1.0.1'))
  "('00000001', '00000000', '00000001', '*final')"

With this change, all 12 test failures I was previously encountering are
now passing. The key to this change is standardizing the old setuptools
output to look just like current setuptools output, using a new function
sanitize_version() in osbs/util.py, which takes output from both variants
of parse_version() and makes it look the same in both cases, so the
results can be operated on with the same code.

A special nod to
http://www.programcreek.com/python/example/7930/pkg_resources.parse_version
for Example 4, which provides the basis for this work-around.

Note: test_build_request.py had duplicate imports for parse_version, and
one of the tests was improperly storing a non-parse_version()'d version,
which tripped up these fixes until remedied.

Includes unit tests for the sanitize_version() function as well, cleaned
up and fully tested on both ancient (el7) and current (f25) setuptools.

Resolves: #564

Signed-off-by: Jarod Wilson <jarod@redhat.com>