build_request: Set 'release' label when provided
Currently if the spec provides a 'release' param, it is not passed along
to the orchestrator as a label.  Detect that and provide it.  Also remove
the bump_release plugin if the 'release' param is passed in.
Updated tests to verify change.

Allow ability to set 'release' number outside of a Dockerfile or a bump_verion()
plugin.  This sets the add_labels_in_dockerfile plugin to be executed by the
orchestrator to set a 'release' label in the Dockerfile passed to the worker.

Untested.  Relying on integration tests to verify this.

[V2] Fix stickler issues
[V3] Add in 'add_labels_in_dockerfile' change

Signed-off-by: Don Zickus <dzickus@redhat.com>