Return Unicode text from log-fetching methods
When streaming or fetching logs, return Unicode text by decoding
the response according to the HTTP headers.

Use response.text to automatically decode the HTTP response when not
streaming it.

When streaming it, response.iter_lines() returns the encoded bytes. Use
the response headers to decide how to decode it; if these are not
present, use guess_json_utf() if JSON is expected, and guess 'utf-8'
otherwise.

We always decode responses before passing them back to the client
because only the response headers say how to interpret the bytes.

Signed-off-by: Tim Waugh <twaugh@redhat.com>