plugins: Move delete_from_registry from worker to orchestrator
For arrangement version 3, the delete_from_registry will move from the worker
node to the orchestrator node.  This is to allow the orchestrator to do some
post-processing of the registries before they are deleted.

In order to make this work successfully, the render_pulp_sync needed to split
out the delete_from_registry pieces into its own render_delete_from_registry().

This would allow the rendering to happen regardless of which node the plugin
actually ran in.  Further some tweaks were added to the render function based
on which arrangement version was being used.  This is for backwards compatibility.

Finally the worker_inner json file for inputs and tests were updated to remove
the delete_from_registry plugin on the worker node in arrangement version 3 as
everything should be working correctly now.