conf: add support for "platform:..." sections
add support to Configuration to parse platform: sections, as described at
http://osbs.readthedocs.io/en/latest/multiarch.html#platform-description.

change get_registry_api_versions to take a platform argument. If there
is a platform argument, only return 'v1' if the configured api version
supports it and that platform value has enable_v1 set to True. Otherwise,
return v2. If the platform doesn't enable v1 and the registry is v1 only,
raise an exception.

Also raise an exception if two or more platforms have enable_v1 set to true.

A stub instance of render_group_manifests() is included, but since
group_manifests isn't in any of the existing arrangements, it will do
nothing for now in practice.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>