pulp_sync: Update delete_from_registry plugin with new args
Moving the delete_from_registry plugin from worker to orchestrator
invovles adding a new arg, 'insecure', to pass in.

The reason is on the worker node, add_docker_registry sets the 'insecure'
flag from another plugin and delete_from_registry just leverages that.

But on the orchestrator, the workflow.push_conf docker registries are
not defined and instead are passed through worker annotations.  Those
annotations do not set 'insecure'.  Change the json file to include
it during 'render'.

The plugin is still backwards compatible with previous arrangement
versions.