Also mount registry_secrets without pulp
When there's no pulp integration, pulp_sync does not run in orchestrator
build. When arrangement_version 4 is used, tag_and_push also does not
run on orchestrator build. In this case, the registry_secrets are never
mounted on build pod. This change ensures that if group_manifests is
enabled, the registry_secrets are also mounted.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>