Make import_image remove dockerRepositoryCheck
The API method import_image is used to force OpenShift to
update the tags tracked in the corresponding ImageStream
object. To do so, the following annotation is cleared:
    "openshift.io/image.dockerRepositoryCheck"
The problem addressed by this commit is that in the case
an ImageStream does not pick up new tags, the API method
may hang indefinitely. Removing this annotation causes
OpenShift to actually "force" the import and reset the
annotation.