render_import_image: fix logic for checking imagestream_name/url
The code was checking for spec.imagestream_name is None not
spec.imagestream_name.value is None - this was never noticed because
spec.imagestream_name is now always set except for Flatpak builds and
previously, a different check was removing the plugin first.

Check in tests that the plugin is properly removed for Flatpak builds.