HttpResponse: better error message when json is corrupt
Currently, if the json string received from a _get is corrupt and
can't be converted, the json package raises a cryptic message about
extra data, but doesn't provide the bad string.

Catch that exception and raise an error with more detail about what
went wrong and the full string that was being loaded.

This doesn't fix osbs-5424, but will hopefully make it easier to
determine what went wrong the next time someone sees a similar failure.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>