api: allow multiple running builds of the same BuildConfig
OpenShift allows multiple builds on the same BuildConfig at the same
time, with only one running and the others queued to run when a build
completes. As such, osbs-client should not raise an error if another
build of the same BuildConfig is active.

Rename _verify_no_running_builds to _verify_running_builds and do not
raise an error if a build is running. Do log an info message of the
names of builds currently running, to make it easier to detect runaway
build processes and other issues that might cause one BuildConfig to
be repeatedly and unnecessarily built.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>