Always include check_and_set_platforms for orchestrated builds
There was a lot of (barely used) complexity in atomic reactor to handle
the check_and_set_platforms plugin being only used for builds with
a koji_target. Including it for all orchestrated builds allows us
to simplify that logic. (This does introduce incompatibility with
older atomic-reactor for builds missing koji_target.)