Increase max time-out and retry backoff factor
Review of the server logs showed that roughly 15% of HTTP requests are
completing after 600 seconds, and almost all the rest are completing
within a few seconds. Our current backoff and retry strategy of starting
retries after 4 seconds is inefficient, and the max time-out of 600 seconds
means that some tasks that complete on the server side are failing on the
osbs-client side because osbs-client is timing out too early.

Increase the back-off factor to 60 seconds, reduce the number of retries
to 60, and increase the max time-out to 1900 seconds so that all the
retries have time to complete. Include documentation explaining this
change.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>