test_arrangements: copy key literals from atomic_reactor.constants
Make it slightly harder to confuse a plugin's name and it's key in
test_arrangements by copying the key literals from atomic_reactor.constants
and using them in test_arrangements.

Doing this will hopefully make it easier to detect and avoid a mismatch
between the plugin name in inputs/*_inner:?.json and the plugins' key
in atomic-reactor.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>