utils: better error message when we fail to parse registry URI
Prior to this change, when a user specified a registry URI that
completely fails to match our validation regex, match() returned None,
and we called .groups() on that None value. This led to the following
error message:

  osbs.exceptions.OsbsException: AttributeError("'NoneType' object has no attribute 'groups'")

Handle the case where our regex does not match the user input and raise
a more helpful error message to the user.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>