user_params: better error for missing ":" in build-from
Prior to this change, when a user specified a --build-from option that
does not contain a ":" character, split() returned one value,
but we tried to assign the result to two values (source_type and
source_value). This led to the following error:

  osbs.exceptions.OsbsException: ValueError('not enough values to unpack (expected 2, got 1)')

Handle the case where the user fails to specify ":" by returning a more
helpful error message to the user.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>