conf: delete support for deprecated keys
fixes OSBS-8383

Now that set_params() isn't calling all the deprecated keys, and the
deprecated keys have been deprecated for several months, get rid of
support for them.

Add a dummy get_deprecated_key() function to preserve the unit test
for deprecated keys so that there is no need to rewrite it if we
end up with deprecated keys in the future.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>