Import unicode_literals in labels module
In cea3df7, we moved the Labels class into its own module. When doing
so, we did not import unicode_literals to the new module. As a
consequence, the keys in the labels dicts became 'str's in python 2.
This patch sets osbs-client back to the former behavior, which is py2/3
consistent.

Signed-off-by: Athos Ribeiro <athos@redhat.com>