Log response text from OCP
Added hook to requests to log response when a HTTP request failed.

However with retries it's tricky, if failure limit is reached, it will
just fail withou passing response to hooks. To be able to process
request while using retries, Retry(raise_on_status=False) has to be
used.

However this param is available from urllib3 > 1.15, in RHEL7 there is
urllib3 ~= 1.10, so response text will be logged only with newer urllib3
version.

* CLOUDBLD-2800

Signed-off-by: Martin Bašti <mbasti@redhat.com>