api.py: Catch mismatches between the flatpak parameter and the repo
If a build is started with the 'flatpak' parameter, but doesn't have a
container.yaml with a 'flatpak' section, or a build is started without
the 'flatpak' parameter, but has a Flatpak container.yaml, the user
would get super-confusing failures, since the repo-parsing part of the code is
independent of the part of the code that checks for build type.

Check for consistency, and throw helpful messages.

Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>