fix task name for the logs
Return the logs with `pipelineTaskName` instead of
task run name. Since, task run name is just
pipeline run name appended with some random
ID it's not of much use. Instead, the pipeline
task name is the one we need since it can help
distinguish the platforms in the logs of specific
tasks.

* CLOUDBLD-9886

Signed-off-by: Harsh Modi <hmodi@redhat.com>