Make ImageName objects immutable
CLOUDBLD-8232

Instances of the ImageName class are currently both hashable and mutable at the same time. This commit makes them immutable and adds a replace() method to allow for changing the object's attributes by returning a copy, not by the in-place mutation. **Code that mutates ImageName objects in place is not changed just yet.**

Signed-off-by: Oliver Lesko <olesko@redhat.com>